core94: Adjust update.sh script if run multiple times
Some changes cannot be executed again after the update
was already installed.

+++ config/rootfiles/core/94/update.sh
@@ -71,14 +71,14 @@ EOF
 fcrontab -z &>/dev/null
 
 # DMA - reconfigure Postfix if exists
-if [ -e /etc/postfix/main.cf ]; then
+if [ -e /etc/postfix/main.cf ] && [ ! -e "/usr/sbin/sendmail.postfix" ]; then
 	mv /usr/sbin/sendmail /usr/sbin/sendmail.postfix
 	/usr/sbin/alternatives --install /usr/sbin/sendmail sendmail /usr/sbin/sendmail.postfix 15
 	sed -i 's/usr\/sbin\/sendmail/usr/sbin/sendmail.postfix/' /opt/pakfire/db/rootfiles/postfix
 fi
 # DMA - configure dma as default mta
-mkdir /etc/alternatives
-mkdir /var/lib/alternatives
+mkdir -p /etc/alternatives
+mkdir -p /var/lib/alternatives
 /usr/sbin/alternatives --install /usr/sbin/sendmail sendmail /usr/sbin/sendmail.dma 20
 
 # Start services
